chore: Rename `validate_inputs.ts` to `validate-inputs.ts` and adjust `if` condition formatting.
