+++ action.yml
@@ -104,7 +104,8 @@ runs:
   using: 'composite'
   steps:
     - name: 'Install pnpm'
-      if: ${{ inputs.use_pnpm == 'true' }}
+      if: |-
+        ${{ inputs.use_pnpm == 'true' }}
       uses: 'pnpm/action-setup@41ff72655975bd51cab0327fa583b6e92b6d3061' # ratchet:pnpm/action-setup@v4
       with:
         version: 10
@@ -124,7 +125,7 @@ runs:
       working-directory: '${{ github.action_path }}'
       shell: 'bash'
       run: |
-        npx ts-node src/validate_inputs.ts
+        npx ts-node src/validate-inputs.ts
       env:
         INPUT_GEMINI_API_KEY: '${{ inputs.gemini_api_key }}'
         INPUT_GOOGLE_API_KEY: '${{ inputs.google_api_key }}'
